+++ runway/data_types.py
@@ -3,17 +3,49 @@ import math
 import base64
 import inspect
 import json
+import os
+import tarfile
 if sys.version_info[0] < 3:
     from cStringIO import StringIO as IO
 else:
     from io import BytesIO as IO
 import numpy as np
 from PIL import Image
-from .utils import is_url, download_to_temp_dir, try_cast_np_scalar
+from .utils import is_url, extract_tarball, try_cast_np_scalar, download_file
 from .exceptions import MissingArgumentError, InvalidArgumentError
 
+class BaseType(object):
+    """An abstract class that defines a base data type interface. This type
+    should be used as the base class of new data types, never directly.
 
-class any(object):
+    :param data_type: The data type represented as a string
+    :type data_type: string
+    :param name: The name associated with this variable, defaults to None
+    :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
+    """
+
+    def __init__(self, data_type, name=None, description=None):
+        self.type = data_type
+        self.name = name
+        self.description = description
+
+    def serialize(self, value):
+        raise NotImplementedError()
+
+    def deserialize(self, value):
+        raise NotImplementedError()
+
+    def to_dict(self):
+        return {
+            'name': self.name,
+            'type': self.type,
+            'description': self.description
+        }
+
+class any(BaseType):
     """A generic data type. The value this data type takes must be serializable to JSON.
 
     .. code-block:: python
@@ -33,10 +65,13 @@ class any(object):
 
     :param name: The name associated with this variable, defaults to "field"
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     """
 
-    def __init__(self, name=None):
-        self.name = name
+    def __init__(self, name='field', description=None):
+        super(any, self).__init__('any', name=name, description=description)
 
     def serialize(self, v):
         return v
@@ -45,12 +80,9 @@ class any(object):
         return v
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'any'
-        ret['name'] = self.name or 'field'
-        return ret
+        return super(any, self).to_dict()
 
-class array(object):
+class array(BaseType):
     """A data type representing an array (list) of other runway.data_type objects.
 
     .. code-block:: python
@@ -68,13 +100,17 @@ class array(object):
     :param name: The name associated with this variable, defaults to "{item_type}_array" \
         (e.g. "image_array")
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param min_length: The minimum number of elements required to be in the array, defaults to 0
     :type min_length: int, optional
     :param max_length: The maximum number of elements allowed to be in the array, defaults to None
     :type max_length: int, optional
     :raises MissingArgumentError: A missing argument error if item_type is not specified
     """
-    def __init__(self, item_type=None, name=None, min_length=0, max_length=None):
+    def __init__(self, item_type=None, name=None, description=None, min_length=0, max_length=None):
+        super(array, self).__init__('array', name=name, description=description)
         if item_type is None: raise MissingArgumentError('item_type')
         if inspect.isclass(item_type):
             self.item_type = item_type()
@@ -91,16 +127,14 @@ class array(object):
         return [self.item_type.serialize(item) for item in items]
 
     def to_dict(self):
-        ret = {}
-        ret['name'] = self.name
-        ret['type'] = 'array'
+        ret = super(array, self).to_dict()
         ret['itemType'] = self.item_type.to_dict()
         ret['minLength'] = self.min_length
         if self.max_length: ret['maxLength'] = self.max_length
         return ret
 
 
-class image(object):
+class image(BaseType):
     """A data type representing an image. Images represent PIL or numpy
     images but are passed to and from the Model SDK as base64 encoded data URI
     strings over the network
@@ -128,6 +162,9 @@ class image(object):
 
     :param name: The name associated with this variable, defaults to "image"
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param channels: The number of channels in the image, defaults to 3. \
         E.g. an "rgb" image has 3 channels while an "rgba" image has 4.
     :type channels: int, optional
@@ -144,8 +181,8 @@ class image(object):
     :param height: The height of the image, defaults to None
     :type height: int, optional
     """
-    def __init__(self, name=None, channels=3, min_width=None, min_height=None, max_width=None, max_height=None, width=None, height=None):
-        self.name = name or 'image'
+    def __init__(self, name='image', description=None, channels=3, min_width=None, min_height=None, max_width=None, max_height=None, width=None, height=None):
+        super(image, self).__init__('image', name=name, description=description)
         self.channels = channels
         self.min_width = min_width
         self.min_height = min_height
@@ -172,9 +209,7 @@ class image(object):
         return 'data:image/jpeg;base64,' + base64.b64encode(buffer.getvalue()).decode('utf8')
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'image'
-        ret['name'] = self.name
+        ret = super(image, self).to_dict()
         ret['channels'] = self.channels
         if self.min_width: ret['minWidth'] = self.min_width
         if self.max_width: ret['maxWidth'] = self.max_width
@@ -185,7 +220,7 @@ class image(object):
         return ret
 
 
-class vector(object):
+class vector(BaseType):
     """A data type representing a vector of floats.
 
     .. code-block:: python
@@ -202,17 +237,21 @@ class vector(object):
             rand = np.random.random_sample(args["length"])
             return { "vector": vec.deserialize(rand) }
 
-    :param length: The number of elements in the vector
-    :type length: int
     :param name: The name associated with this variable, defaults to "vector"
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
+    :param length: The number of elements in the vector
+    :type length: int, inferred if a default vector is specified
     :param sampling_mean: The mean of the sample the vector is drawn from, defaults to 0
     :type sampling_mean: float, optional
     :param sampling_std: The standard deviation of the sample the vector is drawn from, defaults to 1
     :type sampling_std: float, optional
     :raises MissingArgumentError: A missing argument error if length is not specified
     """
-    def __init__(self, length=None, name=None, default=None, sampling_mean=0, sampling_std=1):
+    def __init__(self, name='vector', description=None, length=None, default=None, sampling_mean=0, sampling_std=1):
+        super(vector, self).__init__('vector', name=name, description=description)
         if default is not None:
             if length is None:
                 length = len(default)
@@ -221,7 +260,6 @@ class vector(object):
                 raise InvalidArgumentError(msg)
         if length is None:
             raise MissingArgumentError('length')
-        self.name = name or 'vector'
         self.length = length
         self.sampling_mean = sampling_mean
         self.sampling_std = sampling_std
@@ -234,9 +272,7 @@ class vector(object):
         return value.tolist()
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'vector'
-        ret['name'] = self.name
+        ret = super(vector, self).to_dict()
         ret['length'] = self.length
         ret['samplingMean'] = self.sampling_mean
         ret['samplingStd'] = self.sampling_std
@@ -244,7 +280,7 @@ class vector(object):
         return ret
 
 
-class category(object):
+class category(BaseType):
     """A categorical data type that allows you to specify a variable's value \
         as a member of a set list of choices.
 
@@ -262,6 +298,9 @@ class category(object):
 
         :param name: The name associated with this variable, defaults to "category"
         :type name: string, optional
+        :param description: A description of this variable and how its used in the model,
+            defaults to None
+        :type description: string, optional
         :param choices: A list of categories, defaults to None
         :type choices: A list of strings
         :param default: A default list of categories, defaults to None
@@ -273,33 +312,32 @@ class category(object):
             choices list.
     """
 
-    def __init__(self, name=None, choices=None, default=None):
+    def __init__(self, name='category', description=None, choices=None, default=None):
+        super(category, self).__init__('category', name=name, description=description)
         if choices is None or len(choices) == 0: raise MissingArgumentError('choices')
         if default is not None and default not in choices:
             msg = 'default argument {} is not in choices list'.format(default)
             raise InvalidArgumentError(msg)
-        self.name = name or 'category'
         self.choices = choices
         self.default = default or self.choices[0]
 
     def deserialize(self, value):
         if value not in self.choices:
-            raise InvalidArgumentError(self.name)
+            msg = 'category value "%s" does not appear in choices list.' % value
+            raise InvalidArgumentError(self.name, msg)
         return value
 
     def serialize(self, value):
         return value
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'category'
-        ret['name'] = self.name
+        ret = super(category, self).to_dict()
         ret['oneOf'] = self.choices
         ret['default'] = self.default
         return ret
 
 
-class number(object):
+class number(BaseType):
     """A basic number data type. Instantiate this class to create a new runway model variable.
 
     .. code-block:: python
@@ -313,6 +351,9 @@ class number(object):
 
     :param name: The name associated with this variable, defaults to None
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param default: A default value for this number variable, defaults to 0
     :type default: float, optional
     :param min: The minimum allowed value of this number type, defaults to 0
@@ -325,8 +366,8 @@ class number(object):
     :type step: float, optional
     """
 
-    def __init__(self, name=None, default=0, min=0, max=1, step=1):
-        self.name = name or 'number'
+    def __init__(self, name='number', description=None, default=0, min=0, max=1, step=1):
+        super(number, self).__init__('number', name=name, description=description)
         self.default = default
         self.min = min
         self.max = max
@@ -339,9 +380,7 @@ class number(object):
         return try_cast_np_scalar(value)
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'number'
-        ret['name'] = self.name
+        ret = super(number, self).to_dict()
         ret['default'] = self.default
         ret['min'] = self.min
         ret['max'] = self.max
@@ -349,7 +388,7 @@ class number(object):
         return ret
 
 
-class text(object):
+class text(BaseType):
     """A basic text data type. Used to represent strings. \
         Instantiate this class to create a new runway model variable.
 
@@ -364,16 +403,19 @@ class text(object):
 
     :param name: The name associated with this variable, defaults to "text"
     :type name: string, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
     :param default: The default value for this text variable, defaults to ''
-    :type default: str, optional
+    :type default: string, optional
     :param min_length: The minimum character length of this text variable, defaults to 0
     :type min_length: int, optional
     :param max_length: The maximum character length of this text variable, \
         defaults to None, which allows text to be of any maximum length
     :type max_length: int, optional
     """
-    def __init__(self, name=None, default='', min_length=0, max_length=None):
-        self.name = name or 'text'
+    def __init__(self, name='text', description=None, default='', min_length=0, max_length=None):
+        super(text, self).__init__('text', name=name, description=description)
         self.default = default
         self.min_length = min_length
         self.max_length = max_length
@@ -385,17 +427,15 @@ class text(object):
         return str(value)
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'text'
-        ret['name'] = self.name
+        ret = super(text, self).to_dict()
         ret['default'] = self.default
         ret['minLength'] = self.min_length
         if self.max_length: ret['maxLength'] = self.max_length
         return ret
 
 
-class file(object):
-    """A data type that represents a file or folder. The file can be a local \
+class file(BaseType):
+    """A data type that represents a file or directory. The file can be a local \
         resource on disk or a remote resource loaded over HTTP. \
         Instantiate this class to create a new runway model variable.
 
@@ -404,34 +444,48 @@ class file(object):
         import runway
         from runway.data_types import file, category
 
-        inputs = {"folder": file(is_folder=True)}
+        inputs = {"directory": file(is_directory=True)}
         outputs = {"result": category(choices=["success", "failure"])}
         @runway.command("batch_process", inputs=inputs, outputs=outputs)
         def batch_process(result_of_setup, args):
-            result = do_something_with(args["folder"])
+            result = do_something_with(args["directory"])
             return { "result": "success" if result else "failure" }
 
     :param name: The name of this variable, defaults to None
     :type name: string, optional
-    :param is_folder: Does this variable represent a folder instead of a file? Defaults to False.
-    :type is_folder: bool, optional
+    :param description: A description of this variable and how its used in the model,
+        defaults to None
+    :type description: string, optional
+    :param is_directory: Does this variable represent a directory instead of a file? Defaults to False.
+    :type is_directory: bool, optional
+    :param extension: Accept only files of this extension.
+    :type extension: string, optional
     """
 
-    def __init__(self, name=None, is_folder=False):
-        self.name = name or 'file'
-        self.is_folder = is_folder
+    def __init__(self, name='file', description=None, is_directory=False, extension=None):
+        super(file, self).__init__('file', name=name, description=description)
+        self.is_directory = is_directory
+        self.extension = extension
 
-    def deserialize(self, value):
-        if is_url(value):
-            return download_to_temp_dir(value)
-        return value
+    def deserialize(self, path_or_url):
+        if is_url(path_or_url):
+            downloaded_path = download_file(path_or_url)
+            if tarfile.is_tarfile(downloaded_path):
+                return extract_tarball(downloaded_path)
+            else:
+                return downloaded_path
+        else:
+            if not os.path.exists(path_or_url):
+                raise InvalidArgumentError('file path provided does not exist')
+            if self.extension and not path_or_url.endswith(self.extension):
+                raise InvalidArgumentError('file path does not have expected extension')
+            return path_or_url
 
     def serialize(self, value):
         return value
 
     def to_dict(self):
-        ret = {}
-        ret['type'] = 'file'
-        ret['name'] = self.name
-        if self.is_folder: ret['isFolder'] = self.is_folder
+        ret = super(file, self).to_dict()
+        if self.is_directory: ret['isDirectory'] = self.is_directory
+        if self.extension: ret['extension'] = self.extension
         return ret