+++ tests/test_data_types.py
@@ -18,6 +18,12 @@ def check_data_type_interface(data_type):
     assert callable(data_type.to_dict)
 
 # We arbitrarily use this release tag to test file download and serialization
+def check_expected_contents_for_0057_file_download(path):
+    assert os.path.isfile(path)
+    with open(path, 'r') as f:
+        assert f.read() == '# Runway Python SDK\n'
+
+# We arbitrarily use this release tag to test file download and serialization
 def check_expected_contents_for_0057_tar_download(path):
     readme_path = os.path.join(path, 'model-sdk-0.0.57', 'README.md')
     assert os.path.isfile(readme_path)
@@ -25,6 +31,9 @@ def check_expected_contents_for_0057_tar_download(path):
         assert f.read() == '# Runway Python SDK\n'
 
 # BASIC TESTS FOR ALL DATA TYPES -----------------------------------------------
+def test_data_type_interface_base_type():
+    check_data_type_interface(BaseType)
+
 def test_data_type_interface_any():
     check_data_type_interface(any)
 
@@ -49,16 +58,67 @@ def test_data_type_interface_text():
 def test_data_type_interface_file():
     check_data_type_interface(file)
 
+# BASE TYPE --------------------------------------------------------------------
+def test_base_type_to_dict():
+
+    base_type = BaseType('base', name='No name', description='Some description.')
+    obj = base_type.to_dict()
+    assert obj['type'] == 'base'
+    assert obj['name'] == 'No name'
+    assert obj['description'] == 'Some description.'
+
+# The BaseType is an abstract class that requires its serialize/deserialize
+# methods to be overwritten by subclasses
+def test_base_type_serialize_not_implemented():
+    base_type = BaseType('base')
+    with pytest.raises(NotImplementedError):
+        base_type.serialize('test')
+
+def test_base_type_deserialize_not_implemented():
+    base_type = BaseType('base')
+    with pytest.raises(NotImplementedError):
+        base_type.deserialize('test')
+
+# ANY --------------------------------------------------------------------------
+def test_any_to_dict():
+    a = any()
+    obj = a.to_dict()
+    assert obj['type'] == 'any'
+    assert obj['name'] == 'field'
+    assert obj['description'] == None
+
+def test_any_serialization():
+    a = any()
+    assert a.serialize(512) == 512
+    assert a.serialize(512.5) == 512.5
+    assert a.serialize('512') == '512'
+    assert a.serialize(None) == None
+    assert a.serialize(True) == True
+    assert a.serialize([]) == []
+    assert a.serialize({}) == {}
+
+def test_any_deserialize():
+    a = any()
+    assert a.deserialize(512) == 512
+    assert a.deserialize(512.5) == 512.5
+    assert a.deserialize('512') == '512'
+    assert a.deserialize(None) == None
+    assert a.deserialize(True) == True
+    assert a.deserialize([]) == []
+    assert a.deserialize({}) == {}
+
 # TEXT -------------------------------------------------------------------------
 def test_text_to_dict():
     default = 'Some default text'
-    txt = text(default=default, min_length=1, max_length=20)
+    description = 'A description about this variable.'
+    txt = text(default=default, description=description, min_length=1, max_length=20)
     obj = txt.to_dict()
     assert obj['type'] == 'text'
     assert obj['name'] == 'text'
     assert obj['default'] == default
     assert obj['minLength'] == 1
     assert obj['maxLength'] == 20
+    assert obj['description'] == description
 
 def test_text_serialization():
     txt = text()
@@ -71,7 +131,8 @@ def test_text_deserialize():
 # NUMBER -----------------------------------------------------------------------
 def test_number_to_dict():
     default = 42
-    num = number(default=default, min=10, max=100)
+    description = 'A description about this variable.'
+    num = number(default=default, description=description, min=10, max=100)
     obj = num.to_dict()
     assert obj['type'] == 'number'
     assert obj['name'] == 'number'
@@ -79,6 +140,7 @@ def test_number_to_dict():
     assert obj['min'] == 10
     assert obj['max'] == 100
     assert obj['step'] == 1
+    assert obj['description'] == description
 
 def test_number_serialization():
     assert 1 == number().serialize(1)
@@ -96,13 +158,15 @@ def test_number_serialize_numpy_scalar():
 
 # ARRAY ------------------------------------------------------------------------
 def test_array_to_dict():
-    arr = array(item_type=text, min_length=5, max_length=10)
+    description = 'A description about this variable.'
+    arr = array(item_type=text, description=description, min_length=5, max_length=10)
     obj = arr.to_dict()
     assert obj['name'] == 'text_array'
     assert obj['type'] == 'array'
     assert obj['itemType'] == text().to_dict()
     assert obj['minLength'] == 5
     assert obj['maxLength'] == 10
+    assert obj['description'] == description
 
 def test_array_no_item_type():
     with pytest.raises(MissingArgumentError):
@@ -129,13 +193,15 @@ def test_array_deserialization():
 
 # VECTOR -----------------------------------------------------------------------
 def test_vector_to_dict():
-    vec = vector(length=128, sampling_mean=0, sampling_std=1)
+    description = 'A description about this variable.'
+    vec = vector(length=128, description=description, sampling_mean=0, sampling_std=1)
     obj = vec.to_dict()
     assert obj['name'] == 'vector'
     assert obj['type'] == 'vector'
     assert obj['length'] == 128
     assert obj['samplingMean'] == 0
     assert obj['samplingStd'] == 1
+    assert obj['description'] == description
 
 def test_vector_no_item_type():
     with pytest.raises(MissingArgumentError):
@@ -167,12 +233,14 @@ def test_vector_default_no_length_arg():
 
 # CATEGORY ---------------------------------------------------------------------
 def test_category_to_dict():
-    cat = category(choices=['one', 'two', 'three'], default='two')
+    description = 'A description about this variable.'
+    cat = category(choices=['one', 'two', 'three'], default='two', description=description)
     obj = cat.to_dict()
     assert obj['name'] == 'category'
     assert obj['type'] == 'category'
     assert obj['oneOf'] == ['one', 'two', 'three']
     assert obj['default'] == 'two'
+    assert obj['description'] == description
 
 def test_category_serialization():
     cat = category(choices=['one', 'two', 'three'], default='two')
@@ -213,13 +281,16 @@ def test_file_to_dict():
     obj = f.to_dict()
     assert obj['name'] == 'file'
     assert obj['type'] == 'file'
+    assert obj['description'] == None
 
-def test_file_to_dict_folder():
-    f = file(is_folder=True)
+def test_file_to_dict_directory():
+    description = 'A description about this variable.'
+    f = file(is_directory=True, description=description)
     obj = f.to_dict()
     assert obj['name'] == 'file'
     assert obj['type'] == 'file'
-    assert obj['isFolder'] == True
+    assert obj['isDirectory'] == True
+    assert obj['description'] == description
 
 def test_file_serialization_base():
     f = file()
@@ -227,7 +298,7 @@ def test_file_serialization_base():
 
 def test_file_serialization_relative():
     f = file()
-    assert 'folder/file.txt' == f.serialize('folder/file.txt')
+    assert 'directory/file.txt' == f.serialize('directory/file.txt')
 
 def test_file_serialization_absolute():
     f = file()
@@ -237,63 +308,66 @@ def test_file_serialization_remote():
     f = file()
     url = 'https://github.com/runwayml/model-sdk/archive/0.0.57.tar.gz'
     assert url == f.serialize(url)
-## TODO: accept ftp:// protocol
-#     url = 'ftp://demo:password@test.rebex.net/readme.txt'
-#     assert url == f.serialize(url)
 
-def test_file_serialization_base_folder():
-    f = file(is_folder=True)
-    assert 'folder' == f.serialize('folder')
+def test_file_serialization_base_directory():
+    f = file(is_directory=True)
+    assert 'directory' == f.serialize('directory')
 
-def test_file_serialization_relative_folder():
-    f = file(is_folder=True)
-    assert 'folder/folder' == f.serialize('folder/folder')
+def test_file_serialization_relative_directory():
+    f = file(is_directory=True)
+    assert 'directory/directory' == f.serialize('directory/directory')
 
-def test_file_serialization_absolute_folder():
-    f = file(is_folder=True)
-    assert '/home/user/folder' == f.serialize('/home/user/folder')
+def test_file_serialization_absolute_directory():
+    f = file(is_directory=True)
+    assert '/home/user/directory' == f.serialize('/home/user/directory')
 
-def test_file_serialization_remote_folder():
-    f = file(is_folder=True)
+def test_file_serialization_remote_directory():
+    f = file(is_directory=True)
     url = 'https://github.com/runwayml/model-sdk/archive/0.0.57.tar.gz'
     assert url == f.serialize(url)
-## TODO: accept ftp:// protocol
-#     url = 'ftp://demo:password@test.rebex.net/'
-#     assert url == f.serialize(url)
 
 def test_file_deserialization_base():
     f = file()
-    assert 'file.txt' == f.deserialize('file.txt')
+    assert 'README.md' == f.deserialize('README.md')
 
 def test_file_deserialization_relative():
     f = file()
-    assert 'folder/file.txt' == f.deserialize('folder/file.txt')
+    assert 'runway/__init__.py' == f.deserialize('runway/__init__.py')
 
 def test_file_deserialization_absolute():
+    absolute_path = os.path.abspath('README.md')
     f = file()
-    assert '/home/user/file.txt' == f.deserialize('/home/user/file.txt')
+    assert absolute_path == f.deserialize(absolute_path)
+
+def test_file_deserialization_not_exist():
+    with pytest.raises(InvalidArgumentError):
+        file().deserialize('file-that-does-not-exist.txt')
+
+def test_file_deserialization_invalid_extension():
+    with pytest.raises(InvalidArgumentError):
+        file(extension='.txt').deserialize('README.md')
 
 def test_file_deserialization_remote():
     f = file()
-    url = 'https://github.com/runwayml/model-sdk/archive/0.0.57.tar.gz'
+    url = 'https://raw.githubusercontent.com/runwayml/model-sdk/0.0.57/README.md'
     path = f.deserialize(url)
     assert os.path.exists(path)
-    check_expected_contents_for_0057_tar_download(path)
+    check_expected_contents_for_0057_file_download(path)
 
-def test_file_deserialization_base_folder():
-    f = file(is_folder=True)
-    assert 'folder' == f.deserialize('folder')
+def test_file_deserialization_base_directory():
+    f = file(is_directory=True)
+    assert 'runway' == f.deserialize('runway')
 
-def test_file_deserialization_relative_folder():
-    f = file(is_folder=True)
-    assert 'folder/folder' == f.deserialize('folder/folder')
+def test_file_deserialization_relative_directory():
+    f = file(is_directory=True)
+    assert 'docs/source' == f.deserialize('docs/source')
 
-def test_file_deserialization_absolute_folder():
-    f = file(is_folder=True)
-    assert '/home/user/folder' == f.deserialize('/home/user/folder')
+def test_file_deserialization_absolute_directory():
+    f = file(is_directory=True)
+    assert '/usr/bin' == f.deserialize('/usr/bin')
 
-def test_file_deserialization_remote_folder():
-    f = file(is_folder=True)
+def test_file_deserialization_remote_directory():
+    f = file(is_directory=True)
     url = 'https://github.com/runwayml/model-sdk/archive/0.0.57.tar.gz'
     path = f.deserialize(url)
     assert os.path.exists(path)
@@ -310,10 +384,11 @@ def test_image_to_dict():
     assert obj['maxWidth'] == 512
     assert obj['minHeight'] == 128
     assert obj['maxHeight'] == 512
+    assert obj['description'] == None
 
 def test_image_serialize_and_deserialize():
-    folder = os.path.dirname(os.path.realpath(__file__))
-    img = Image.open(os.path.join(folder, 'test_image.jpg'))
+    directory = os.path.dirname(os.path.realpath(__file__))
+    img = Image.open(os.path.join(directory, 'test_image.jpg'))
     serialized_pil = image().serialize(img)
     deserialized_pil = image().deserialize(serialized_pil)
     assert issubclass(type(deserialized_pil), Image.Image)