+++ tests/test_model.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # Ensure that the local version of the runway module is used, not a pip
 # installed version
 import sys
@@ -7,11 +8,14 @@ sys.path.insert(0, '.')
 import os
 import json
 import pytest
+from time import sleep
 from runway.model import RunwayModel
+from runway.__version__ import __version__ as model_sdk_version
 from runway.data_types import category, text, number, array, image, vector, file, any as any_type
 from runway.exceptions import *
-from utils import get_test_client
+from utils import get_test_client, get_manifest
 from deepdiff import DeepDiff
+from flask import abort
 
 os.environ['RW_NO_SERVE'] = '1'
 
@@ -23,23 +27,31 @@ def test_model_setup_and_command():
     closure = dict(setup_ran = False, command_ran = False)
 
     expected_manifest = {
+        'modelSDKVersion': model_sdk_version,
+        'millisRunning': None,
+        'millisSinceLastCommand': None,
+        'GPU': os.environ.get('GPU', False),
         'options': [{
             'type': 'category',
             'name': 'size',
             'oneOf': ['big', 'small'],
-            'default': 'big'
+            'default': 'big',
+            'description': 'The size of the model. Bigger is better but also slower.',
         }],
         'commands': [{
             'name': 'test_command',
+            'description': None,
             'inputs': [{
                 'type': 'text',
                 'name': 'input',
+                'description': 'Some input text.',
                 'default': '',
                 'minLength': 0
             }],
             'outputs': [{
                 'type': 'number',
                 'name': 'output',
+                'description': 'An output number.',
                 'default': 0,
                 'min': 0,
                 'max': 1,
@@ -50,14 +62,24 @@ def test_model_setup_and_command():
 
     rw = RunwayModel()
 
-    @rw.setup(options={ 'size': category(choices=['big', 'small']) })
+    description = 'The size of the model. Bigger is better but also slower.'
+    @rw.setup(options={ 'size': category(choices=['big', 'small'], description=description) })
     def setup(opts):
         closure['setup_ran'] = True
         return {}
 
-    inputs = { 'input': text }
-    outputs = { 'output': number }
-    @rw.command('test_command', inputs=inputs, outputs=outputs)
+    inputs = { 'input': text(description='Some input text.') }
+    outputs = { 'output': number(description='An output number.') }
+
+    # Python 2.7 doesn't seem to handle emoji serialization correctly in JSON,
+    # so we will only test emoji serialization/deserialization in Python 3
+    if sys.version_info[0] < 3:
+        description = 'Sorry, Python 2 doesn\'t support emoji very well'
+    else:
+        description = 'A test command whose description contains emoji 🕳'
+    expected_manifest['commands'][0]['description'] = description
+
+    @rw.command('test_command', inputs=inputs, outputs=outputs, description=description)
     def test_command(model, opts):
         closure['command_ran'] = True
         return 100
@@ -66,13 +88,37 @@ def test_model_setup_and_command():
 
     client = get_test_client(rw)
 
-    # check the manifest via a GET /
-    response = client.get('/')
+    response = client.get('/meta')
+    assert response.is_json
+
     manifest = json.loads(response.data)
+
+    # unset millisRunning as we can't reliably predict this value.
+    # testing that it is an int should be good enough.
+    assert type(manifest['millisRunning']) == int
+    manifest['millisRunning'] = None
+
     assert manifest == expected_manifest
+
+    # TEMPORARILY CHECK / PATH IN ADDITION TO /meta ----------------------------
+    # ... sorry for the gross dupe code ;)
+    response = client.get('/')
+    assert response.is_json
+
+    manifest = json.loads(response.data)
+
+    # unset millisRunning as we can't reliably predict this value.
+    # testing that it is an int should be good enough.
+    assert type(manifest['millisRunning']) == int
+    manifest['millisRunning'] = None
+
+    assert manifest == expected_manifest
+    # --------------------------------------------------------------------------
 
     # check the input/output manifest for GET /test_command
     response = client.get('/test_command')
+    assert response.is_json
+
     command_manifest = json.loads(response.data)
     assert command_manifest == expected_manifest['commands'][0]
 
@@ -80,7 +126,13 @@ def test_model_setup_and_command():
         'input': 'test input'
     }
     response = client.post('/test_command', json=post_data)
+    assert response.is_json
     assert json.loads(response.data) == { 'output' : 100 }
+
+    # now that we've run a command lets make sure millis since last command is
+    # a number
+    manifest_after_command = get_manifest(client)
+    assert type(manifest_after_command['millisSinceLastCommand']) == int
 
     assert closure['command_ran'] == True
     assert closure['setup_ran'] == True
@@ -96,7 +148,8 @@ def test_model_healthcheck():
     rw.run(debug=True)
     client = get_test_client(rw)
     response = client.get('/healthcheck')
-    assert response.data == b'RUNNING'
+    assert response.is_json
+    assert response.json == { 'status': 'RUNNING' }
 
 def test_model_setup_no_arguments():
 
@@ -144,6 +197,50 @@ def test_model_options_missing():
         with pytest.raises(MissingOptionError):
             rw.run(debug=True)
 
+def test_setup_invalid_category():
+
+    rw = RunwayModel()
+    @rw.setup(options={'category': category(choices=['Starks', 'Lannisters'])})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/setup', json={ 'category': 'Tyrells' })
+
+    assert response.status_code == 400
+    json_response = json.loads(response.data)
+    assert 'error' in json_response
+    # ensure the user is displayed an error that indicates the category option
+    # is problematic
+    assert 'Invalid argument: category' in json_response['error']
+    # ensure the user is displayed an error that indicates the problematic value
+    assert 'Tyrells' in json_response['error']
+
+def test_command_invalid_category():
+
+    rw = RunwayModel()
+    inputs = {'category': category(choices=['Starks', 'Lannisters'])}
+    outputs = {'reflect': text }
+    @rw.command('test_command', inputs=inputs, outputs=outputs)
+    def test_command(opts):
+        return opts['category']
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/test_command', json={ 'category': 'Targaryen' })
+
+    assert response.status_code == 400
+    json_response = json.loads(response.data)
+    assert 'error' in json_response
+    # ensure the user is displayed an error that indicates the category option
+    # is problematic
+    assert 'Invalid argument: category' in json_response['error']
+    # ensure the user is displayed an error that indicates the problematic value
+    assert 'Targaryen' in json_response['error']
+
 def test_meta(capsys):
 
     rw = RunwayModel()
@@ -166,6 +263,7 @@ def test_meta(capsys):
         pass
 
     kwargs_2 = {
+        'description': 'This command is used for testing.',
         'inputs': {
             'any': any_type,
             'file': file
@@ -185,25 +283,30 @@ def test_meta(capsys):
                 'minLength': 0,
                 'type': 'array',
                 'name': 'initialization_array',
+                'description': None,
                 'itemType': {
                     'default': '',
                     'minLength': 0,
                     'type': 'text',
-                    'name': 'text'
+                    'name': 'text',
+                    'description': None
                 }
             }
         ],
         'commands': [
             {
                 'name': 'command_2',
+                'description': 'This command is used for testing.',
                 'inputs': [
                     {
                         'type': 'any',
                         'name': 'any',
+                        'description': None,
                     },
                     {
                         'type': 'file',
                         'name': 'file',
+                        'description': None,
                     },
                 ],
                 'outputs': [
@@ -214,16 +317,19 @@ def test_meta(capsys):
                         'max': 100,
                         'step': 1,
                         'type': 'number',
+                        'description': None
                     },
                 ]
             },
             {
                 'name': 'command_1',
+                'description': None,
                 'inputs': [
                     {
                         'channels': 3,
                         'type': 'image',
                         'name': 'image',
+                        'description': None
                     },
                     {
                         'samplingMean': 0,
@@ -231,7 +337,8 @@ def test_meta(capsys):
                         'type': 'vector',
                         'name': 'vector',
                         'samplingStd': 1,
-                        'default': [0, 0, 0, 0, 0]
+                        'default': [0, 0, 0, 0, 0],
+                        'description': None
                     },
                 ],
                 'outputs': [
@@ -239,7 +346,8 @@ def test_meta(capsys):
                         'default': '',
                         'minLength': 0,
                         'type': 'text',
-                        'name': 'label'
+                        'name': 'label',
+                        'description': None
                     }
                 ]
             }
@@ -262,6 +370,213 @@ def test_meta(capsys):
     assert std.err == ''
 
     os.environ['RW_META'] = '0'
+
+def test_post_setup_json_no_mime_type():
+
+    rw = RunwayModel()
+
+    @rw.setup(options={'input': text})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/setup', data='{"input": "test input"}')
+    assert response.is_json
+    assert json.loads(response.data) == { 'success': True }
+
+def test_post_setup_invalid_json_no_mime_type():
+
+    rw = RunwayModel()
+
+    @rw.setup(options={'input': text})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/setup', data='{"input": test input"}')
+
+    assert response.is_json
+    assert response.status_code == 400
+
+    expect = { 'error': 'The body of all POST requests must contain JSON' }
+    assert json.loads(response.data) == expect
+
+
+def test_post_setup_json_mime_type():
+
+    rw = RunwayModel()
+
+    @rw.setup(options={'input': text})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/setup', json={ 'input': 'test input' })
+    assert response.is_json
+    assert json.loads(response.data) == { 'success': True }
+
+def test_post_setup_form_encoding():
+
+    rw = RunwayModel()
+
+    @rw.setup(options={'input': text})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    content_type='application/x-www-form-urlencoded'
+    response = client.post('/setup', data='input=test', content_type=content_type)
+
+    assert response.is_json
+    assert response.status_code == 400
+
+    expect = { 'error': 'The body of all POST requests must contain JSON' }
+    assert json.loads(response.data) == expect
+
+def test_post_command_json_no_mime_type():
+
+    rw = RunwayModel()
+
+    @rw.command('times_two', inputs={ 'input': number }, outputs={ 'output': number })
+    def times_two(model, args):
+        return args['input'] * 2
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/times_two', data='{ "input": 5 }')
+    assert response.is_json
+    assert json.loads(response.data) == { 'output': 10 }
+
+def test_post_command_json_mime_type():
+
+    rw = RunwayModel()
+
+    @rw.command('times_two', inputs={ 'input': number }, outputs={ 'output': number })
+    def times_two(model, args):
+        return args['input'] * 2
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.post('/times_two', json={ 'input': 5 })
+    assert response.is_json
+    assert json.loads(response.data) == { 'output': 10 }
+
+def test_post_command_form_encoding():
+
+    rw = RunwayModel()
+
+    @rw.command('times_two', inputs={ 'input': number }, outputs={ 'output': number })
+    def times_two(model, args):
+        return args['input'] * 2
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    content_type='application/x-www-form-urlencoded'
+    response = client.post('/times_two', data='input=5', content_type=content_type)
+    assert response.is_json
+    assert response.status_code == 400
+
+    expect = { 'error': 'The body of all POST requests must contain JSON' }
+    assert json.loads(response.data) == expect
+
+def test_405_method_not_allowed():
+
+    rw = RunwayModel()
+
+    @rw.setup(options={'input': text})
+    def setup(opts):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.put('/setup', json= { 'input': 'test input'})
+
+    assert response.is_json
+    assert response.status_code == 405
+    assert response.json == { 'error': 'Method not allowed.' }
+
+def test_404_not_found():
+
+    rw = RunwayModel()
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.get('/asfd')
+
+    assert response.is_json
+    assert response.status_code == 404
+
+def test_401_unauthorized():
+
+    rw = RunwayModel()
+
+    @rw.app.route('/test/unauthorized')
+    def unauthorized():
+        abort(401)
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.get('/test/unauthorized')
+
+    assert response.is_json
+    assert response.status_code == 401
+
+    expect = { 'error': 'Unauthorized (well... really unauthenticated but hey I didn\'t write the spec).' }
+    assert response.json == expect
+
+def test_403_forbidden():
+
+    rw = RunwayModel()
+
+    @rw.app.route('/test/forbidden')
+    def unauthorized():
+        abort(403)
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.get('/test/forbidden')
+
+    assert response.is_json
+    assert response.status_code == 403
+
+    expect = { 'error': 'Forbidden.' }
+    assert response.json == expect
+
+def test_500_internal_server_error():
+
+    rw = RunwayModel()
+
+    @rw.app.route('/test/internal_server_error')
+    def unauthorized():
+        abort(500)
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+    response = client.get('/test/internal_server_error')
+
+    assert response.is_json
+    assert response.status_code == 500
+
+    expect = { 'error': 'Internal server error.' }
+    assert response.json == expect
 
 def test_setup_error_setup_no_args():
 
@@ -300,4 +615,76 @@ def test_inference_error():
     rw.run(debug=True)
 
     response = client.post('test_command', json={ 'input': 5 })
+    assert response.is_json
     assert 'InferenceError' in str(response.data)
+
+def test_millis_since_run_increases_over_time():
+
+    rw = RunwayModel()
+    client = get_test_client(rw)
+    rw.run(debug=True)
+
+    last_time = get_manifest(client)['millisRunning']
+    assert type(last_time) == int
+    for i in range(3):
+        sleep(0.01)
+        millis_running = get_manifest(client)['millisRunning']
+        assert millis_running > last_time
+        last_time = millis_running
+
+def test_millis_since_last_command_resets_each_command():
+
+    rw = RunwayModel()
+
+    @rw.command('test_command', inputs={ 'input': number }, outputs = { 'output': text })
+    def test_command(model, inputs):
+        pass
+
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    assert get_manifest(client)['millisSinceLastCommand'] is None
+    client.post('test_command', json={ 'input': 5 })
+
+    first_time = get_manifest(client)['millisSinceLastCommand']
+    assert type(first_time) == int
+
+    for i in range(5):
+        sleep(0.02)
+        millis_since_last_command = get_manifest(client)['millisSinceLastCommand']
+        assert millis_since_last_command > first_time
+        client.post('test_command', json={ 'input': 5 })
+        assert get_manifest(client)['millisSinceLastCommand'] < millis_since_last_command
+
+def test_gpu_in_manifest_no_env_set():
+
+    rw = RunwayModel()
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    if os.environ.get('GPU') is not None:
+        del os.environ['GPU']
+
+    assert get_manifest(client)['GPU'] == False
+
+def test_gpu_in_manifest_gpu_env_true():
+
+    rw = RunwayModel()
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    os.environ['GPU'] = '1'
+    assert get_manifest(client)['GPU'] == True
+
+def test_gpu_in_manifest_gpu_env_false():
+
+    rw = RunwayModel()
+    rw.run(debug=True)
+
+    client = get_test_client(rw)
+
+    os.environ['GPU'] = '0'
+    assert get_manifest(client)['GPU'] == False